uuid v.3

+++ src/main/java/com/digneequipe/hardoize/services/AuthService.java
@@ -117,7 +117,7 @@ public class AuthService {
                 .email(u.getEmail())
                 .role(u.getRole())
                 .photoUri(u.getPhotoUri())
-                .evaluation(u.getEvaluation())
+                .evaluation(u.getEvaluation().floatValue())
                 .estActif(u.getEstActif())
                 .createdAt(u.getCreatedAt())
                 .build();
